[fe#59] rename: 폴더 이름 변경
