+++ fe/src/common/productCard/eventWrapper/eventWrapper.js
