+++ fe/src/components/header/gnbLists/gnbLists.styled.js
