+++ fe/src/components/header/svg/svg.styled.js
