+++ fe/src/components/main/bestProduct/tabList/tabList.styled.js
