+++ fe/src/components/main/categoryArea/CategoryArea.js
