+++ fe/src/components/main/categoryArea/category/category.styled.js
